Move ChromeTracing in its own file, and add tracing support for VR server (App_SharedMemoryPhysics_VR) Add a bit of extra sleep in PhysicsServer thread, to make rendering smoother.
