+++ examples/Utils/ChromeTraceUtil.h
@@ -1,0 +1,9 @@
+
+#ifndef B3_CHROME_TRACE_UTIL_H
+#define B3_CHROME_TRACE_UTIL_H
+
+void b3ChromeUtilsStartTimings();
+void b3ChromeUtilsStopTimingsAndWriteJsonFile();
+void b3ChromeUtilsEnableProfiling();
+
+#endif//B3_CHROME_TRACE_UTIL_H
